refactor: extract StreamingTTSManager to modules/speech/streaming.py (#546)
Phase 4.1 — move StreamingTTSManager class (220 lines) from orchestrator.py
to its own module. numpy import made lazy (inside _cache_full_audio method)
to avoid import errors in cloud mode. Unused orchestrator imports removed
(hashlib, re, OrderedDict, ThreadPoolExecutor, numpy).

orchestrator.py: 4287 → 4062 lines (−225)
